Correctly parse numbers with scientific notation
The parser split on 'e' and 'E' as if it were an instruction, which for
'1e1' or '1e-1' would introduce an extra numeric argument, and of course
result in the wrong value for the first argument.
The solution is to leave 'e' and 'E' untouched.

In addition to that '-' was always prefixed with a space, which again
causes problems with '1e-1'. A solution for this is to only insert
spaces if the '-' is not preceded by an 'e' or 'E'.